+++ core-service/src/main/java/com/mka/webmarket/core/entities/Order.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -28,7 +29,7 @@ public class Order {
     @Column(name = "phone")
     private String phone;
     @Column(name = "total_price")
-    private int totalPrice;
+    private BigDecimal totalPrice;
     @CreationTimestamp
     @Column(name = "created_at")
     private LocalDateTime createdAt;
